refactored auth folder structure and its mutation and query js files
