--- modules/auth/resolvers/login.js
@@ -1,25 +1,0 @@
-const bcrypt = require("bcrypt");
-const jwt = require("jsonwebtoken");
-
-const User = require("../../../models/User");
-
-const login = async (_, { loginInput: { email, password } }) => {
-  const user = await User.findOne({ email: email });
-  if (!user) {
-    throw new Error("User does not exist!");
-  }
-  const isEqual = await bcrypt.compare(password, user.password);
-  if (!isEqual) {
-    throw new Error("Password is incorrect!");
-  }
-  const token = jwt.sign(
-    { userId: user.id, email: user.email },
-    "somesupersecretkey",
-    {
-      expiresIn: "1h",
-    }
-  );
-  return { userId: user.id, token: token, tokenExpiration: 1 };
-};
-
-module.exports = login;
