+++ modules/auth/resolvers/queries.js
@@ -1,0 +1,5 @@
+const User = require("../../../models/User");
+
+
+
+module.exports = {}
