+++ modules/index.js
@@ -1,6 +1,6 @@
 const { makeExecutableSchemaFromModules } = require('../utils/modules')
 
-const auth = require('./auth')
+const auth = require('./auth/typeDefs')
 const users = require('./users')
 
 module.exports = makeExecutableSchemaFromModules({
